requirements.txt
fastapi==0.115.4
google_api_python_client==2.149.0
protobuf==5.28.3
python-dotenv==1.0.1
Requests==2.32.3